Change: Move api.js into util/
